requirements.txt
altair==4.2.2
asn1crypto==1.5.1
attrs==23.1.0
backports.zoneinfo==0.2.1
bcrypt==4.0.1
blinker==1.6.2
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
cryptography==40.0.2
decorator==5.1.1
entrypoints==0.4
extra-streamlit-components==0.1.56
filelock==3.12.0
gitdb==4.0.10
GitPython==3.1.31
idna==3.4
importlib-metadata==6.6.0
importlib-resources==5.12.0
Jinja2==3.1.2
jsonschema==4.17.3
markdown-it-py==2.2.0
MarkupSafe==2.1.2
mdurl==0.1.2
numpy==1.24.3
oscrypto==1.3.0
packaging==23.1
pandas==2.0.1
Pillow==9.5.0
pkgutil_resolve_name==1.3.10
protobuf==3.20.3
pyarrow==10.0.1
pycparser==2.21
pycryptodomex==3.17
pydeck==0.8.1b0
Pygments==2.15.1
PyJWT==2.6.0
Pympler==1.0.1
pyOpenSSL==23.1.1
pyrsistent==0.19.3
python-dateutil==2.8.2
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
requests==2.29.0
rich==13.3.5
six==1.16.0
smmap==5.0.0
snowflake-connector-python==3.0.3
streamlit==1.22.0
streamlit-authenticator==0.2.1
tenacity==8.2.2
toml==0.10.2
toolz==0.12.0
tornado==6.3.1
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==4.3
urllib3==1.26.15
validators==0.20.0
watchdog==3.0.0
zipp==3.15.0
